safari-534.53-branch/LayoutTests/fast/files/file-reader-abort-expected.txt

CONSOLE MESSAGE: line 18: ReferenceError: Can't find variable: FileReader
FAIL: Timed out waiting for notifyDone to be called

Test that FileReader.abort works.
